9/24: - 상태변경에 따른 애니매이션 출력 - 애니매이션적용

+++ CScene_Start.cpp
@@ -92,10 +92,10 @@ void CScene_Start::render(HDC _dc)
 	Vec2 vRenderPos = CCamera::GetInst()->GetRenderPos(m_vForcePos);
 
 	::Ellipse(_dc
-		, vRenderPos.x - m_fCurRadius
-		, vRenderPos.y - m_fCurRadius
-		, vRenderPos.x + m_fCurRadius
-		, vRenderPos.y + m_fCurRadius);
+		, (int)(vRenderPos.x - m_fCurRadius)
+		, (int)(vRenderPos.y - m_fCurRadius)
+		, (int)(vRenderPos.x + m_fCurRadius)
+		, (int)(vRenderPos.y + m_fCurRadius));
 }
 
 void CScene_Start::Enter()
